move query planning and execution into scheduler and executor modules
